+++ src/modules/persistence/users.persistence.service.ts
@@ -9,7 +9,7 @@ import { IUsersPersistenceService } from './interfaces/users.persistence.service
 import { TimeSlotDbModel } from '../../models/timeslot.db.model';
 import { UserDbModel, UserRole, PasswordStatus } from '../../models/user.db.model';
 import { UserUniqueValidationRequest } from '../../models/user-unique-validation-request.model';
-import { EmailBody, sendemail } from '../../utils/nodeMailer';
+import { EmailBody, sendemail } from '../../Utils/node-mailer';
 
 @Injectable()
 export class UsersPersistenceService implements IUsersPersistenceService {
